static casting uint to int for safety

+++ src/simulation_system/Engine.hpp
@@ -98,9 +98,9 @@ class Engine
 
     // eigen parallelization parameters
     // number of physical CPU cores on host device
-    const int m_num_physical_cores{std::thread::hardware_concurrency()};
+    const int m_num_physical_cores{static_cast<int>(std::thread::hardware_concurrency())};
     /// @todo: number of physical CPU cores to use in tensor calculations
-    const int m_simulation_cores{std::ceil(m_num_physical_cores / 4)};
+    const int m_simulation_cores{static_cast<int>(std::ceil(m_num_physical_cores / 4))};
 
     // ProgressBar output
     /// Percentage of simulation progress at which to output a GSD frame
